Better view updating
Move the view update to the provision step. This is more in line with
how we would want to use this in a pipeline for building artifacts
with plans. It will require the views exist for each of the delivery
stages.

We also need to track some information about the artifact that was
built in a data bag so that information can be used elsewhere in the
pipeline - such as for the view update.

Signed-off-by: jtimberman <joshua@chef.io>