Use delivery-secrets for origin key
We should load the origin keys from a delivery-secrets data bag item.
This commit introduces a helper method to determine whether that is
available and that all the required data is present. It doesn't
validate the content, just that it exists and is non-nil.

Otherwise, we continue to generate the key with the hab command.

Then the key files are written out to the target locations, one in the
build node, and a copy in the studio.