Refactor databag/environment to publish action
Clean up the publish recipe by removing the unnecessary local
variables (build_version is now a method, we only use project_name in
one place).

Move the data bag and environment code into the publish action of the
hab_build resource. This allows us to have more flexibility, and the
publish action can be used on multiple project packages built in a
custom build cookbook.

Use the new resource name for the data bag and its item. Make sure to
`lazy` the values because they're calculated using the helper methods
for loading the last build data.

Consolidate the multiline command to make it easier to read, and make
rubocop happier about the length of the block.

Signed-off-by: Joshua Timberman <joshua@chef.io>