fix helper methods not present in resource
Commit 87e29ad moved the defined helper methods from a global scope into
a module. However, they were not "sent" to the `Chef::Resource`, which
caused errors when using the `hab_build` resource. This commit fixes the
problems by sending the DSL methods to `Chef::Resource`, and also
defines the methods that are only needed inside the resource in its
`action_class` block.

We also have fixed up the tests. The test kitchen suites will run
through a "publish" phase use case where we test the `hab_build`
resource is added to the resource collection as we expect it to be.
We've added ChefSpec to test this. While we won't be able to upload, we
do have that action available. The plan we build works in test kitchen
to build an artifact. The test cookbook's default recipe will set up the
test kitchen node so it can run the `delivery` cli job, which should more
closely mimic what we'll see on Chef Automate.

Signed-off-by: Nathan L Smith <smith@chef.io>
Signed-off-by: jtimberman <joshua@chef.io>