Save application release as separate action
This change moves saving an application release
to a separate action. By moving this out to its own
action we will allow users to control when the
when the release is saved. This is needed for builds
that need to do extra work after the habitat build is
completed.

Note: This change is non-backwards compatable for those
using the hab_build resource's publish action directly.

Signed-off-by: Jon Morrow <jmorrow@chef.io>